// ui-packages/ui-components/lib/sidenav/sidenav.type.ts
export type SidenavStandardLibrary = LibrariesItem_Response;
// ...
export enum SidenavItemId {
  Files = 'Files',
  Examples = 'Examples',
  Libraries = 'Libraries',
  Reference = 'Reference',
  GenAI = 'GenAI',
  Settings = 'Settings',
}
// ...
export type GetLibrary = (
  params: GetLibrary_Params,
  enabled: boolean,
) => {
  library?: Omit<LibraryDetails_Response, 'examples'> & {
    examples?: Example[];
  };
  isLoading: boolean;
  refetch: () => void;
};
// ...
export interface ExampleFile {
  name: string;
  path: string;
}

export interface Example {
  name: string;
  path: string;
  folder?: string;
  types?: string[];
  ino?: ExampleFile;
  files?: ExampleFile[];
}

export interface ExamplesFolder {
  name: string;
  examples: ExampleItem[];
  examplesNumber: number;
}

export type CustomLibraryExamplesFolder = {
  name: string;
  path: string;
  examplesNumber: number;
};

export type ExampleItem = Example | ExamplesFolder;
// ...
export enum SidenavExamplesIds {
  Device = 'Device',
  BuiltIn = 'BuiltIn',
  FromLibraries = 'FromLibraries',
}